requirements.txt
feedparser>=6.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
trafilatura>=2.0.0
openai>=1.0.0
